return top 20

+++ Trakx.MarketData.Feeds.Common/Models/CoinMarketCap/IQuote.cs
@@ -4,12 +4,12 @@ namespace Trakx.MarketData.Feeds.Common.Models.CoinMarketCap
 {
     public interface IQuote
     {
-        double Price { get; }
+        decimal Price { get; }
         double Volume24H { get; }
         double PercentChange1H { get; }
         double PercentChange24H { get; }
         double PercentChange7D { get; }
-        double MarketCap { get; }
+        decimal MarketCap { get; }
         DateTimeOffset LastUpdated { get; }
     }
 }
